refactor API client
